bug fix: _occupied_indices now size_t vector. not uint32_t

+++ src/lib/share/mempool_hash_table_st.h
@@ -29,7 +29,7 @@ private:
     // 如果 node 存在非平凡析构对象, 那么对 HashTableNode 显式调用析构
     void destroy_node(HashTableNode* node) {
         if constexpr (!std::is_trivially_destructible<HashTableNode>::value) {
-            node->~HashTableNode();
+            node->~HashTableNode(); // HashTableNode 没有显式定义 析构函数, 编译器会生成默认的析构函数: 其会逆序递归析构所有需要析构的成员
         }
     }
 
@@ -65,8 +65,8 @@ private:
     // gc 链表: 链起所有node
     HashTableNode* _all_nodes_head = nullptr;
 
-    // 记录非空bucket index. 桶置空操作时只需遍历这些桶即可. index用uint32_t就够了，因为它可代表42亿大的hashtable
-    std::vector<uint32_t> _occupied_indices;
+    // 记录非空bucket index. 桶置空操作时只需遍历这些桶即可. index类型要与 _capacity 类型对齐, 因为它是 hash成员函数的输出 取_capacity余
+    std::vector<size_t> _occupied_indices;
 
     // 指针传入内存池（模板方式传入。用纯虚类接口的方式传入过不了编译器）
     TYPE_MEMPOOL* _pool; // void* allocate(size_t size)
@@ -91,7 +91,7 @@ private:
         if (!_new_table) throw std::bad_alloc();
 
         // 新的非空桶列表
-        std::vector<uint32_t> _new_occupied_indices;
+        std::vector<size_t> _new_occupied_indices;
         _new_occupied_indices.reserve(_occupied_indices.size());
 
         // 新的gc链表
@@ -101,7 +101,7 @@ private:
         size_t actual_node_count = 0;
 
         // 遍历非空桶
-        for (uint32_t old_index: _occupied_indices) {
+        for (size_t old_index: _occupied_indices) {
 
             HashTableNode* curr = _table[old_index];
             while (curr) { // 当前node非空
@@ -112,7 +112,7 @@ private:
                 curr->next = _new_table[new_index]; // 当前node挂载到新bucket链表头
                 _new_table[new_index] = curr; // 更新确认新bucket的链表头
                 // 若空桶->非空, 记录
-                if (was_empty) _new_occupied_indices.push_back(static_cast<uint32_t>(new_index));
+                if (was_empty) _new_occupied_indices.push_back(new_index);
                 // 头插到gc链表
                 curr->gc_next = _new_all_nodes_head;
                 _new_all_nodes_head = curr;
@@ -151,7 +151,7 @@ public:
         alloc_table_ptrs(_capacity);
     }
 
-    // 析构函数, 会调用 destroy 方法来释放所有 HashTableNode 中需要显式析构的部分, 释放 buckets数组 _table并置空, _size 和 _capacity 置0
+    // 析构函数, 会调用 destroy 方法来 析构 所有 HashTableNode 中需要显式析构的部分, 释放 buckets数组 _table并置空, _size 和 _capacity 置0
     // 但不负责内存释放. 由内存池在外部统一释放
     ~hash_table_st_chain() {
         destroy();
@@ -212,7 +212,7 @@ public:
         new_node->gc_next = _all_nodes_head;
 
         // 如果 空 -> 非空, 那么记录桶号
-        if (was_empty) _occupied_indices.push_back(static_cast<uint32_t>(index));
+        if (was_empty) _occupied_indices.push_back(index);
 
         // node数量自加1. 原子线程安全
         ++_size;
@@ -252,7 +252,7 @@ public:
 
         new_node->gc_next = _all_nodes_head;
 
-        if (was_empty) _occupied_indices.push_back(static_cast<uint32_t>(index));
+        if (was_empty) _occupied_indices.push_back(index);
 
         _size++;
 
@@ -264,9 +264,9 @@ public:
     }
 
 
-    // 哈希表是构建在传入的 内存池 上的数据结构, 它不应该负责 内存池 的销毁
+    // 哈希表是构建在传入的 内存池 上的数据结构, 它不应该负责 内存池 的 复位or销毁
     // 内存池本身是只可以 整体复用/整体销毁，不可精确销毁单次allocate的内存
-    // 哈希表的"清空"：原数据全部析构, 不再可访问, 但其分配的内存不会在这里被销毁. 保持 bucket 结构
+    // 哈希表的"清空"：原数据全部析构, 不再可访问, 但其分配的内存不会在这里被 复位or销毁. 保持 bucket 结构
     // 由于保持了 bucket 结构 和 内存池, 故 reset 内存池之后, 本哈希表即可重新复用(insert/upsert node)
     void clear() {
         if (_capacity == 0 || !_table) {
@@ -276,7 +276,7 @@ public:
             return;
         }
 
-        if constexpr(!std::is_trivially_destructible<HashTableNode>::value) {
+        if constexpr(!std::is_trivially_destructible<HashTableNode>::value) { // constexpr: 编译时求值
             // 若 node 需要非平凡析构：沿着 gc 链 析构所有node
             HashTableNode* curr = _all_nodes_head;
             while (curr) {
@@ -289,11 +289,11 @@ public:
         // gc 链表置空
         _all_nodes_head = nullptr;
 
-        // 非空桶置空. _table 指针数组保持结构
-        for (uint32_t index: _occupied_indices) {
+        // _table 指针数组保持结构
+        for (size_t index: _occupied_indices) {
             _table[index] = nullptr;
         }
-
+        // 非空桶置空
         _occupied_indices.clear();
         _size = 0;
 
@@ -322,8 +322,10 @@ public:
         // gc 链表置空
         _all_nodes_head = nullptr;
 
-        // 释放 节点指针数组, 置空 _table
+        // 释放 节点指针(桶)数组, 置空 _table. 所有桶/节点不再可访问. 但内存尚未释放, 等待内存池操作
         free_table_ptrs();
+
+        // 非空桶置空
         _occupied_indices.clear();
         _size = 0; // node数量置0
         _capacity = 0; // _capacity 置零
